% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTCN.R
\name{testCn}
\alias{testCn}
\title{Edf Test for Poisson Distribution \eqn{Cn}}
\usage{
testCn(x, n.boot)
}
\arguments{
\item{x}{vector of nonnegative integers, the sample data}

\item{n.boot}{number of bootstrap replicates}
}
\description{
Performs the empirical distribution function goodness-of-fit test of Poisson distribution with unknown parameter
}
\value{The function \code{testCn} returns a list with class \code{htest} containing:
\item{}{Description of test}
\item{data}{Description of data}
\item{test statistic}{Value of test statistic}
\item{p-value}{approximate p-value of the test}
\item{mean}{sample mean}
}

\details{The edf test of Poissonity \eqn{Cn} was proposed by Henze (1996). The test is based on the similarity between the edf of the random variable \eqn{X} and the cdf of a Poisson distribution with parameter mean of \eqn{X}. The test statistic is a Cramer-von-Mises type of distance.

The test is implemented by parametric boostrap with \code{n.boot} replicates}

\author{Manuel Mendez Hurtado \email{mmendezinformatica@gmail.com}}

\references{Henze, N. (1996) Empirical-distribution-function goodness-of-fit tests for discrete models, \emph{The Canadian Journal of Statistics} Vol 24 No 1, 81-93 \url{https://www.jstor.org/stable/3315691?seq=1}}

\examples{x <- rpois(20,2)
testCn(x, n.boot = 500)}
