% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTCN.R
\name{testCn_}
\alias{testCn_}
\title{Edf Test for Poisson Distribution \eqn{Cn*}}
\usage{
testCn_(x, n.boot)
}
\arguments{
\item{x}{vector of nonnegative integers, the sample data}

\item{n.boot}{number of bootstrap replicates}
}
\description{
Performs the empirical distribution function goodness-of-fit test of Poisson distribution with unknown parameter
}
\value{The function \code{testCn_} returns a list with class \code{htest} containing:
\item{}{Description of test}
\item{data}{Description of data}
\item{test statistic}{Value of test statistic}
\item{p-value}{approximate p-value of the test}
\item{mean}{sample mean}
}

\details{The edf test of Poissonity \eqn{Cn*} was proposed by Henze (1996). The test is based on the similarity between the edf of the random variable \eqn{X}, \eqn{F_{n}(k)}, and the cdf of Poisson distribution with parameter \eqn{\lambda = \hat{X}}, \eqn{F(k,\hat{\lambda}_{n})}. The test statistic is a modification of the Cramer-von-Mises type of distance.

\eqn{C_{n}^\star = n \sum_{k = 0}^{\infty} [F_{n}(k) - F(k;\hat{\lambda}_{n})]^2 f_{n}(k)}

\eqn{f_{n}(k)} denotes \eqn{F_{n}(k) - F_{n}(k-1)}.
The test is implemented by parametric boostrap with \code{n.boot} replicates}

\author{Manuel Mendez Hurtado \email{mmendezinformatica@gmail.com}}

\references{Henze, N. (1996) Empirical-distribution-function goodness-of-fit tests for discrete models, \emph{The Canadian Journal of Statistics} Vol 24 No 1, 81-93 \url{https://www.jstor.org/stable/3315691?seq=1}}

\examples{x <- rpois(20,2)
testCn_(x, n.boot = 500)}
