% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTCN.R
\name{testdeltainf}
\alias{testdeltainf}
\title{Pgf Test for Poisson Distribution \eqn{\Delta_\infty}}
\usage{
testdeltainf(x, n.boot)
}
\arguments{
\item{x}{vector of nonnegative integers, the sample data}

\item{n.boot}{number of bootstrap replicates}
}
\description{
Performs the empirical distribution function goodness-of-fit test of Poisson distribution with unknown parameter
}
\value{The function \code{testdeltainf} returns a list with class \code{htest} containing:
\item{}{Description of test}
\item{data}{Description of data}
\item{test statistic}{Value of test statistic}
\item{p-value}{approximate p-value of the test}
\item{mean}{sample mean}
}

\details{The test of Poissonity \eqn{\Delta_{\infty}} was proposed by Puig and Weibb (2020). The test is based on some characterizations like the Binomial thinning operator for the probability generating function of a Poisson distribution. The test statistic is based on the \eqn{L \infty} norm.

\eqn{\hat{\Delta}^\star_{\infty}=\max_{t \in [0,1]} (|\hat{g}(t) - [\hat{g}(t)(\frac{t+1}{2})]^2|)}

\eqn{\hat{g}(t)} denotes the empirical pobability generating function of X.         
         
The test is implemented by parametric boostrap with \code{n.boot} replicates}

\author{Manuel Mendez Hurtado \email{mmendezinformatica@gmail.com}}

\references{Puig, P and Weibb, C. H. (2020) Empirical-distribution-function goodness-of-fit tests for discrete models, \emph{Computational Statistics and Data Analysis} Vol 144, 1-12 \url{https://www.sciencedirect.com/science/article/pii/S0167947319302336}}

\examples{x <- rpois(20,2)
testdeltainf(x, n.boot = 500)}
