% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTCN.R
\name{testLn}
\alias{testLn}
\title{Edf Test for Poisson Distribution \eqn{Ln}}
\usage{
testLn(x, n.boot)
}
\arguments{
\item{x}{vector of nonnegative integers, the sample data}

\item{n.boot}{number of bootstrap replicates}
}
\description{
Performs the empirical distribution function goodness-of-fit test of Poisson distribution with unknown parameter
}
\value{The function \code{testLn} returns a list with class \code{htest} containing:
\item{}{Description of test}
\item{data}{Description of data}
\item{test statistic}{Value of test statistic}
\item{p-value}{approximate p-value of the test}
\item{mean}{sample mean}
}

\details{The edf test of Poissonity \eqn{Cn} was proposed by Klar (2000). The test is based on the similarity between the edf of the random variable \eqn{X}, \eqn{F_{n}(k)}, and the cdf of a Poisson distribution with parameter \eqn{\lambda = \hat{X}}, \eqn{F(k,\hat{\lambda}_{n})}. The test statistic is created working in the \eqn{l1} space of all sequences satisfying \eqn{\sum{k \ge 0} |X_{k}| < \infty.}

\eqn{L_{n} = \sum_{k \ge 0} \sqrt{n} |F_{n}(k) - F(k,\hat{\lambda}_{n})|}
         
The test is implemented by parametric boostrap with \code{n.boot} replicates}

\author{Manuel Mendez Hurtado \email{mmendezinformatica@gmail.com}}

\references{Klar, B. (2000) Goodness-of-fit tests for discrete models based on the integrated distribution function, \emph{Metrika} Vol 49, 53-69 \url{https://link.springer.com/article/10.1007/s001840050025}}

\examples{x <- rpois(20,2)
testLn(x, n.boot = 500)}
