% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTCN.R
\name{testRn}
\alias{testRn}
\title{Pgf Test for Poisson Distribution \eqn{Rn}}
\usage{
testRn(x, n.boot)
}
\arguments{
\item{x}{vector of nonnegative integers, the sample data.}

\item{n.boot}{number of bootstrap replicates.}
}
\description{
    Performs the probability-generating function goodness-of-fit test of Poisson distribution with unknown parameter.
}
\value{The function \code{testRn} returns a list with class \code{htest} containing:
\item{}{Description of test.}
\item{data}{Description of data.}
\item{test statistic}{Value of test statistic.}
\item{p-value}{approximate p-value of the test.}
\item{mean}{sample mean.}
}

\details{The pgf test of Poissonity \eqn{Rn} was proposed by Rueda et al (1991). The test is based on the similarity between the empirical probability-generating function of the random variable \eqn{X} and the pgf of a Poisson distribution with parameter mean of \eqn{X}. The test statistic is a square similarity measure.

    The test is implemented by parametric boostrap with \code{n.boot} replicates.
  }

\author{Manuel Mendez Hurtado \email{mmendezinformatica@gmail.com}}

\references{Rueda, R., Perez-Abreu, v., O'Reilly, F. (1991) Goodness of fit for the poisson distribution based on the probability generating function \emph{Commun. Statist. - Theory Mehods} Vol 20 (10), 3093-3110 \url{https://www.tandfonline.com/doi/abs/10.1080/03610929108830690}}

\examples{x <- rpois(20,2)
testRn(x, n.boot = 500)
}
