% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTCN.R
\name{testRna}
\alias{testRna}
\title{Pgf Test for Poisson Distribution \eqn{Rn}}
\usage{
testRna(x, a, n.boot)
}
\arguments{
\item{x}{vector of nonnegative integers, the sample data.}

\item{n.boot}{number of bootstrap replicates.}
\item{a}{weight value}
}
\description{
    Performs a weigthed probability-generating function goodness-of-fit test of Poisson distribution with unknown parameter.
}
\value{The function \code{testRna} returns a list with class \code{htest} containing:
\item{}{Description of test.}
\item{data}{Description of data.}
\item{test statistic}{Value of test statistic.}
\item{p-value}{approximate p-value of the test.}
\item{mean}{sample mean.}
}

\details{The pgf test of Poissonity \eqn{Rna} was proposed by Baringhaus et al (2000). The test is based on the similarity between the empirical probability-generating function of the random variable \eqn{X} and the pgf of a Poisson distribution with parameter mean of \eqn{X}. The test statistic is a weighted Cramer-von-Mises distance.

    The test is implemented by parametric boostrap with \code{n.boot} replicates.
  }

\author{Manuel Mendez Hurtado \email{mmendezinformatica@gmail.com}}

\references{Baringhaus, L., Gurtel, N., Henze, N. (2000) Weighted Integral Test Statistics and Components of Smooth Tests of Fit \emph{Austral. New Zeal. J. Statist} \url{ https://doi.org/10.1111/1467-842X.00117}}

\examples{x <- rpois(20,2)
testRna(x, a = 3, n.boot = 500)
}
