% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTCN.R
\name{testTn}
\alias{testTn}
\title{Pgf Test for Poisson Distribution \eqn{Tn}}
\usage{
testTn(x, n.boot)
}
\arguments{
\item{x}{vector of nonnegative integers, the sample data.}
\item{n.boot}{number of bootstrap replicates.}
}
\description{
    Performs the probability-generating function goodness-of-fit test of Poisson distribution with unknown parameter.
}
\value{The function \code{testTn} returns a list with class \code{htest} containing:
\item{}{Description of test.}
\item{data}{Description of data.}
\item{test statistic}{Value of test statistic.}
\item{p-value}{approximate p-value of the test.}
\item{mean}{sample mean.}
}

\details{The pgf test of Poissonity \eqn{Tn} was proposed by Baringhaus and Henze (1992). The test is based on the similarity between the empirical probability-generating function of the random variable \eqn{X} and the pgf of a Poisson distribution with parameter mean of \eqn{X}. The test statistic is a square similarity measure.

\eqn{T_{n} = \int_{0}^{1} [\bar{X}_{n} g_{n}(s) - g`_{n}(s)]^2 ds}

\eqn{g_{n}(s)} denotes pgf of X.         
         
    The test is implemented by parametric boostrap with \code{n.boot} replicates.
  }

\author{Manuel Mendez Hurtado \email{mmendezinformatica@gmail.com}}

\references{Baringhaus, L., Henze, N. (1992) A goodness of fit test for the Poisson distribution based on the empirical generating function \emph{Statist. Probab. Lett.} Vol 13, 269-274 \url{https://ideas.repec.org/a/eee/stapro/v13y1992i4p269-274.html} }

\examples{x <- rpois(20,2)
testTn(x, n.boot = 500)
}
