% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTCN.R
\name{testu}
\alias{testu}
\title{Dispersion Index Test for Poisson Distribution }
\usage{
testu(x, n.boot)
}
\arguments{
\item{x}{vector of nonnegative integers, the sample data.}

\item{n.boot}{number of bootstrap replicates. If \code{n.boot = 0}, it uses the normal approximation method.}
}
\description{
Performs the disperssion index goodness-of-fit test of Poisson distribution with unknown parameter.
}
\value{The function \code{testu} returns a list with class \code{htest} containing:
\item{}{Description of test.}
\item{data}{Description of data.}
\item{test statistic}{Value of test statistic.}
\item{p-value}{approximate p-value of the test.}
\item{mean}{sample mean.}
}
\details{The test of Poissonity \eqn{u} is based on a normalization of the Disperssion Index, equal to 1 in a Poisson Distribution.

\eqn{u = (D \sqrt{\frac{n - 1}{2(1 - 1/N)}}}

\eqn{D} denotes the disperssion index of X.

\eqn{N} denotes the sum of X.

\eqn{n} denotes the length of X.
         
         
The test is implemented by parametric boostrap with \code{n.boot} replicates or by normal approximation.
}

\author{Manuel Mendez Hurtado \email{mmendezinformatica@gmail.com}}


\references{IAEA (2011) Cytogenetic Dosimetry: Applications in Preparedness for and Response to Radiation Emergencies \url{https://www.iaea.org/es/publications/10456/cytogenetic-dosimetry-applications-in-preparedness-for-and-response-to-radiation-emergencies}

Rao, C. R., Chakravarti, I.M. (1956) Some small sample test of significance for a Poisson Distribution \emph{Biometrics} Vol 12, 264-282 \url{https://www.jstor.org/stable/3001466?seq=1}}

\examples{x <- rpois(20,2)
testu(x, n.boot = 500)
testu(x, n.boot = 0)}
